Safari-609.1.11.1/LayoutTests/fast/table/table-create-tbody-multiple-tbody.html

<!DOCTYPE html>
<script src="../../resources/dump-as-markup.js"></script>
<p>https://bugs.webkit.org/show_bug.cgi?id=84465 : Implement createTBody for table element.</p>
<p>This test verify that if a table has multiple tbodies, createTBody() will add the new tbody after the last one.</p>
<table>
  <thead>
  </thead>
  <tfoot>
  </tfoot>
  <tbody title="tbody1">
  </tbody>
  <tbody title="tbody2">
  </tbody>
</table>
<script>
var table = document.getElementsByTagName("table")[0];
table.createTBody();
Markup.dump(table);
</script>
