Safari-611.1.10/LayoutTests/fast/table/table-create-tbody-multiple-tbody-expected.txt

| "\n  "
| <thead>
|   "\n  "
| "\n  "
| <tfoot>
|   "\n  "
| "\n  "
| <tbody>
|   title="tbody1"
|   "\n  "
| "\n  "
| <tbody>
|   title="tbody2"
|   "\n  "
| <tbody>
| "\n"
